refactor: move grid dungeon generation
